pyproject.toml
[project]
name = "faker-ka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
]
